backend/requirements.txt
praw==7.8.1
pandas==2.3.1
numpy==2.3.2
nltk==3.9.1
yfinance==0.2.65
python-dotenv==1.1.1
werkzeug==3.1.3
flask==3.1.1
yfinance==0.2.65
flask-cors==6.0.1
gunicorn==23.0.0
flask-cors==6.0.1
gunicorn==23.0.0
mysql-connector-python=9.5.0
datetime=6.0
gradio_client=2.0.2